I've fixed the duplicate menu in the LoginFragment and added the navigation methods.
To-Do:
I still need to update the menu buttons. For some reason, I can't navigate back to the HomePage using the menu in both the LoginFragment and the RegisterFragment.
Additionally, in the Introduction and Basics XML files, there’s a topic that overlaps with the overlay and obscures some of the text. This needs to be removed.